Core: XmlNode size optimized (40% less average), removed 64bit warning in InVector
git-svn-id: svn://ultimatepp.org/upp/trunk@6151 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/InVector.h
@@ -39,7 +39,7 @@ private:
 	uint16 slave;
 
 	InVectorSlave__ *Slave()          { return (InVectorSlave__ *)((byte *)this + slave); }
-	void SetSlave(InVectorSlave__ *s) { slave = (byte *)s - (byte *)this; }
+	void SetSlave(InVectorSlave__ *s) { slave = (uint16)((byte *)s - (byte *)this); }
 
 	void SetCache(int blki, int offset) const;
 	void ClearCache() const;
